requirements.txt
build==1.2.2.post1
fqdn==1.5.1
furo==2024.8.6
gdal==3.10.1
id==1.5.0
isoduration==20.11.0
jaraco.collections==5.1.0
jsonpointer==3.0.0
matplotlib==3.10.1
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
odfpy==1.4.1
openpyxl==3.1.5
orjson==3.10.15
pandas==2.2.3
pdfkit==1.0.0
pip-chill==1.0.3
psycopg2==2.9.10
pyarrow==19.0.1
pyodbc==5.2.0
pypandoc==1.15
pyproj==3.7.1
pytest-cov==6.0.0
python-rapidjson==1.20
rapidfuzz==3.12.2
scikit-learn==1.6.1
shapely==2.0.7
sphinx-copybutton==0.5.2
sphinx-new-tab-link==0.7.0
sphinx-toggleprompt==0.6.0
sqlalchemy==2.0.39
tinycss2==1.4.0
tomli==2.0.1
twine==6.0.1
ujson==5.10.0
uri-template==1.3.0
webcolors==24.11.1
xlsx2csv==0.8.4
xlsxwriter==3.2.2
